Second commit

+++ src/app/components/laboratoire/laboratoire.component.ts
@@ -15,8 +15,7 @@ export class LaboratoireComponent implements OnInit {
   laboratoires: Laboratoire[];
   etablissements: Etablissement[];
   newLaboratoireForm: FormGroup;
-
-  // laboratoire!: Laboratoire;
+  editLaboratoireForm: FormGroup;
 
   constructor(
     private laboratoireService: LaboratoireService,
@@ -36,6 +35,12 @@ export class LaboratoireComponent implements OnInit {
       departement: ['', Validators.required],
       etablissement: ['', Validators.required]
     });
+    this.editLaboratoireForm = this.formBuilder.group({
+      id: ['',Validators.required],
+      intitule: ['',Validators.required],
+      departement: ['',Validators.required],
+      etablissement: ['', Validators.required]
+    })
   }
 
   getAllLaboratoires(): void {
@@ -77,6 +82,7 @@ export class LaboratoireComponent implements OnInit {
           console.log('New Laboratoire:', newLaboratoire);
           // Reset the form
           this.newLaboratoireForm.reset();
+          this.getAllLaboratoires();
         },
         (error) => {
           console.log(error);
@@ -85,10 +91,22 @@ export class LaboratoireComponent implements OnInit {
     }
   }
 
-  updateLaboratoire(laboratoire: Laboratoire): void {
+  updateLaboratoire(): void {
+    const laboratoire: Laboratoire = {
+      id: this.editLaboratoireForm.value.id,
+      intitule: this.editLaboratoireForm.value.intitule,
+      departement: this.editLaboratoireForm.value.departement,
+      etablissement: {
+        id: this.editLaboratoireForm.value.etablissement.id,
+        intitule: '',
+        adresse: ''
+      }
+    }
     this.laboratoireService.updateLaboratoire(laboratoire).subscribe(
       (updatedLaboratoire: Laboratoire) => {
         // Handle success, if needed
+        this.getAllLaboratoires()
+        console.log(updatedLaboratoire)
       },
       (error: any) => {
         console.error(error);
@@ -97,9 +115,12 @@ export class LaboratoireComponent implements OnInit {
   }
 
   deleteLaboratoire(id: number): void {
+    if (confirm("Etes vous sure !!"))
     this.laboratoireService.deleteLaboratoire(id).subscribe(
-      () => {
+      (r) => {
         // Handle success, if needed
+        console.log(r)
+        this.getAllLaboratoires()
       },
       (error: any) => {
         console.error(error);
@@ -107,4 +128,12 @@ export class LaboratoireComponent implements OnInit {
     );
   }
 
+  public loadObjectData(laboratoire: Laboratoire) {
+    this.editLaboratoireForm.patchValue({
+      id: laboratoire.id,
+      intitule: laboratoire.intitule,
+      departement: laboratoire.departement,
+      etablissement: laboratoire.etablissement
+    });
+  }
 }
